HARDWARE_Blackbox/requirements.txt
opencv-python==4.9.0.80
pyserial==3.5
pyttsx3==2.90
SpeechRecognition==3.10.1
pyaudio==0.2.14
numpy==1.26.4